pyproject.toml
[project]
name = "emoji-haikus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.0",
    "google-cloud-storage>=3.1.0",
    "google-cloud-texttospeech>=2.26.0",
    "google-genai>=1.15.0",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.5",
]
